+++ web/client/src/components/project/ProjectBurndownChart.tsx
@@ -10,7 +10,6 @@ import {
   YAxis,
 } from 'recharts';
 import { ProjectExpenditureProgress } from '@/components/project/ProjectExpenditureProgress.tsx';
-import { ProjectProjectionProgress } from '@/components/project/ProjectProjectionProgress.tsx';
 import { formatCurrency } from '@/lib/currency.ts';
 import {
   PROJECT_PERSONNEL_COLOR,
@@ -310,14 +309,6 @@ export function ProjectBurndownSection({
 
   return (
     <>
-      {projectionQuery.isSuccess && result && series.length > 0 && (
-        <ProjectProjectionProgress
-          awardEndDate={awardEndDate}
-          awardStartDate={awardStartDate}
-          categories={result.categories}
-        />
-      )}
-
       <section className="mt-8 pb-4">
         <h2 className="h2 mb-1">Project Burndown</h2>
         <p className="max-w-3xl mb-6">{tooltipDefinitions.projectBurndown}</p>
@@ -518,7 +509,11 @@ export function ProjectBurndownSection({
       </section>
 
       {projectionQuery.isSuccess && result && series.length > 0 && (
-        <ProjectExpenditureProgress categories={result.categories} />
+        <ProjectExpenditureProgress
+          awardEndDate={awardEndDate}
+          awardStartDate={awardStartDate}
+          categories={result.categories}
+        />
       )}
     </>
   );
